refactor: moved components and split up in fragments
