+++ web/src/components/author-list/author-list.module.css
@@ -1,0 +1,47 @@
+@import '../../styles/custom-properties.css';
+
+.root {
+  margin: 2rem 0 3rem;
+  border-top: 1px solid var(--color-very-light-gray);
+}
+
+.headline {
+  composes: base from '../typography/typography.module.css';
+  font-weight: 600;
+  margin: 0.5rem 0 0;
+}
+
+.list {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+.listItem {
+  font-size: var(--font-small-size);
+  margin: 1rem 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  @nest & > div:last-child {
+    flex: 1;
+    margin-left: 0.5rem;
+  }
+}
+
+.avatar {
+  position: relative;
+  width: 3em;
+  height: 3em;
+  background: var(--color-very-light-gray);
+  border-radius: 50%;
+  overflow: hidden;
+
+  @nest & img {
+    width: 100%;
+    height: 100%;
+    vertical-align: top;
+    object-fit: cover;
+  }
+}
