--- web/src/components/blog-post-preview-grid.module.css
@@ -1,51 +1,0 @@
-@import '../styles/custom-media.css';
-@import '../styles/custom-properties.css';
-
-.root {
-  margin: 2em 0 4em;
-}
-
-.headline {
-  font-size: var(--font-micro-size);
-  line-height: var(--font-micro-line-height);
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-  margin: 2rem 0;
-}
-
-.grid {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-gap: 2em;
-
-  @media (--media-min-small) {
-    grid-template-columns: 1fr 1fr;
-  }
-
-  @media (--media-min-medium) {
-    grid-template-columns: 1fr 1fr 1fr;
-  }
-}
-
-.browseMoreNav {
-  composes: small from './typography.module.css';
-  margin-top: 1rem;
-  text-align: right;
-
-  @nest & a {
-    display: inline-block;
-    padding: 0.5rem 0;
-    color: inherit;
-    text-decoration: none;
-
-    @media (hover: hover) {
-      @nest &:hover {
-        color: var(--color-accent);
-      }
-    }
-  }
-}
