+++ web/src/components/blog-post-preview/blog-post-preview.js
@@ -1,0 +1,45 @@
+import {format} from 'date-fns'
+import {Link} from 'gatsby'
+import React from 'react'
+import {buildImageObj, cn, getBlogUrl} from '../../lib/helpers'
+import {imageUrlFor} from '../../lib/image-url'
+import PortableText from '../portable-text/portable-text'
+
+import styles from './blog-post-preview.module.css'
+import {responsiveTitle3} from '../typography/typography.module.css'
+import {useCurrentLocale} from '../../hooks/i18n'
+
+function BlogPostPreview (props) {
+  const locale = useCurrentLocale()
+
+  return (
+    <Link
+      className={props.isInList ? styles.inList : styles.inGrid}
+      to={getBlogUrl(props.publishedAt, props.slugLocale.current, locale)}
+    >
+      <div className={styles.leadMediaThumb}>
+        {props.mainImage && props.mainImage.asset && (
+          <img
+            src={imageUrlFor(buildImageObj(props.mainImage))
+              .width(600)
+              .height(Math.floor((9 / 16) * 600))
+              .auto('format')
+              .url()}
+            alt={props.mainImage.alt}
+          />
+        )}
+      </div>
+      <div className={styles.text}>
+        <h3 className={cn(responsiveTitle3, styles.title)}>{props.title}</h3>
+        {props._rawExcerpt && (
+          <div className={styles.excerpt}>
+            <PortableText blocks={props._rawExcerpt} />
+          </div>
+        )}
+        <div className={styles.date}>{format(props.publishedAt, 'MMMM Do, YYYY')}</div>
+      </div>
+    </Link>
+  )
+}
+
+export default BlogPostPreview
