+++ web/src/components/blog-post/blog-post.module.css
@@ -1,0 +1,89 @@
+@import '../../styles/custom-media.css';
+@import '../../styles/custom-properties.css';
+
+.root {}
+
+.title {
+  composes: responsiveTitle1 from '../typography/typography.module.css';
+}
+
+.mainImage {
+  display: block;
+  position: relative;
+  background: var(--color-very-light-gray);
+  padding-bottom: calc(9 / 16 * 100%);
+
+  @nest & img {
+    display: block;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    vertical-align: top;
+    object-fit: cover;
+  }
+}
+
+.grid {
+  @media (--media-min-medium) {
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-column-gap: 2em;
+    grid-template-columns: 3fr 1fr;
+  }
+}
+
+.mainContent {
+  @nest & a {
+    color: var(--color-accent);
+
+    @media (hover: hover) {
+      @nest &:hover {
+        color: inherit;
+      }
+    }
+  }
+
+  @nest & h2, & h3, & h4, & h5, & h6 {
+    font-weight: 600;
+  }
+
+  @nest & figure {
+    margin: 0;
+    padding: 0;
+
+    @nest & img {
+      max-width: 100%;
+    }
+  }
+}
+
+.metaContent {
+}
+
+.publishedAt {
+  composes: small from '../typography/typography.module.css';
+  margin: 2rem 0 3rem;
+  color: var(--color-gray);
+}
+
+.categories {
+  border-top: 1px solid var(--color-very-light-gray);
+  margin: 2rem 0 3rem;
+
+  @nest & ul {
+    list-style: none;
+    margin: 0.75rem 0;
+    padding: 0;
+  }
+
+  @nest & ul li {
+    padding: 0.25rem 0;
+  }
+}
+
+.categoriesHeadline {
+  composes: base from '../typography/typography.module.css';
+  margin: 0.5rem 0 0;
+}
