--- web/src/components/container.module.css
@@ -1,12 +1,0 @@
-@import '../styles/custom-media.css';
-
-.root {
-  box-sizing: border-box;
-  max-width: 960px;
-  padding: 1.5em;
-  margin: 0 auto;
-
-  @media (--media-min-small) {
-    padding: 2em;
-  }
-}
