+++ web/src/components/container/container.module.css
@@ -1,0 +1,12 @@
+@import '../../styles/custom-media.css';
+
+.root {
+  box-sizing: border-box;
+  max-width: 960px;
+  padding: 1.5em;
+  margin: 0 auto;
+
+  @media (--media-min-small) {
+    padding: 2em;
+  }
+}
