+++ web/src/components/portable-text/portable-text.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import clientConfig from '../../../client-config'
+import BasePortableText from '@sanity/block-content-to-react'
+import serializers from '../serializers/serializers'
+
+const PortableText = ({blocks}) => (
+  <BasePortableText blocks={blocks} serializers={serializers} {...clientConfig.sanity} />
+)
+
+export default PortableText
