+++ web/src/components/serializers/serializers.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import Figure from '../figure/figure'
+
+const serializers = {
+  types: {
+    authorReference: ({node}) => <span>{node.author.name}</span>,
+    mainImage: Figure
+  }
+}
+
+export default serializers
