--- web/src/components/typography.module.css
@@ -1,138 +1,0 @@
-@import '../styles/custom-media.css';
-@import '../styles/custom-properties.css';
-
-
-
-/*
- * Statically sized elements
- */
-
-.title1 {
-  font-size: var(--font-title1-size);
-  line-height: var(--font-title1-line-height);
-}
-
-.title2 {
-  font-size: var(--font-title2-size);
-  line-height: var(--font-title2-line-height);
-}
-
-.title3 {
-  font-size: var(--font-title3-size);
-  line-height: var(--font-title3-line-height);
-}
-
-.large {
-  font-size: var(--font-large-size);
-  line-height: var(--font-large-line-height);
-}
-
-.base {
-  font-size: inherit;
-  line-height: inherit;
-}
-
-.small {
-  font-size: var(--font-small-size);
-  line-height: var(--font-small-line-height);
-}
-
-.micro {
-  font-size: var(--font-micro-size);
-  line-height: var(--font-micro-line-height);
-  text-transform: uppercase;
-}
-
-
-
-/*
- * Responsively sized elements
- */
-
-.paragraph {
-  font-size: var(--font-base-size);
-  line-height: var(--font-base-line-height);
-  margin: 0.5rem 0 1rem 0;
-
-  @media (--media-min-small) {
-    font-size: var(--font-base-size);
-    line-height: var(--font-base-line-height);
-  }
-
-  @media (--media-min-medium) {
-    font-size: var(--font-large-size);
-    line-height: var(--font-large-line-height);
-  }
-}
-
-.blockQuote {
-  background: var(--color-very-light-gray);
-}
-
-.responsiveTitle1 {
-  font-weight: 600;
-  font-size: var(--font-title2-size);
-  line-height: var(--font-title2-line-height);
-  margin: 1rem 0 2rem;
-
-  @media (--media-min-small) {
-    font-size: var(--font-title2-size);
-    line-height: var(--font-title2-line-height);
-  }
-
-  @media (--media-min-medium) {
-    font-size: var(--font-title1-size);
-    line-height: var(--font-title1-line-height);
-  }
-}
-
-.responsiveTitle2 {
-  font-weight: 600;
-  font-size: var(--font-large-size);
-  line-height: var(--font-large-line-height);
-  margin: 1.5rem 0 0.5rem;
-
-  @media (--media-min-small) {
-    font-size: var(--font-title3-size);
-    line-height: var(--font-title3-line-height);
-  }
-
-  @media (--media-min-medium) {
-    font-size: var(--font-title2-size);
-    line-height: var(--font-title2-line-height);
-  }
-}
-
-.responsiveTitle3 {
-  font-weight: 600;
-  font-size: var(--font-large-size);
-  line-height: var(--font-large-line-height);
-  margin: 1rem 0 0.5rem;
-
-  @media (--media-min-small) {
-    font-size: var(--font-large-size);
-    line-height: var(--font-large-line-height);
-  }
-
-  @media (--media-min-medium) {
-    font-size: var(--font-title3-size);
-    line-height: var(--font-title3-line-height);
-  }
-}
-
-.responsiveTitle4 {
-  font-weight: 600;
-  font-size: var(--font-base-size);
-  line-height: var(--font-base-line-height);
-  margin: 1rem 0 0.5rem;
-
-  @media (--media-min-small) {
-    font-size: var(--font-base-size);
-    line-height: var(--font-base-line-height);
-  }
-
-  @media (--media-min-medium) {
-    font-size: var(--font-large-size);
-    line-height: var(--font-large-line-height);
-  }
-}
