+++ web/src/containers/layout.js
@@ -1,11 +1,11 @@
 import {graphql, StaticQuery} from 'gatsby'
 import React, {useState} from 'react'
-import Layout from '../components/layout'
+import DefaultLayout from '../layouts/default-layout/default-layout'
 
 const query = graphql`
   query SiteTitleQuery {
     site: sanitySiteSettings(_id: {regex: "/(drafts.|)siteSettings/"}) {
-      title
+      ...siteSettings
     }
   }
 `
@@ -28,7 +28,7 @@ function LayoutContainer (props) {
           )
         }
         return (
-          <Layout
+          <DefaultLayout
             {...props}
             showNav={showNav}
             siteTitle={data.site.title}
