+++ web/src/layouts/default-layout/default-layout.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import Header from '../../components/header/header'
+
+import '../../styles/layout.css'
+import styles from './default-layout.module.css'
+
+const DefaultLayout = ({children, onHideNav, onShowNav, showNav, siteTitle}) => (
+  <>
+    <Header siteTitle={siteTitle} onHideNav={onHideNav} onShowNav={onShowNav} showNav={showNav} />
+    <div className={styles.content}>{children}</div>
+    <footer className={styles.footer}>
+      <div className={styles.footerWrapper}>
+        <div className={styles.siteInfo}>
+          &copy; {new Date().getFullYear()}, Built with <a href='https://www.sanity.io'>Sanity</a>{' '}
+          &amp;
+          {` `}
+          <a href='https://www.gatsbyjs.org'>Gatsby</a>
+        </div>
+      </div>
+    </footer>
+  </>
+)
+
+export default DefaultLayout
